Dashboard landing page (#10003)
* Introduce dashboard landing page

Still TODO:
 - Add clickable breadcrumbs
 - Remove New and Open top nav options

* use clickable breadcrumbs instead of 'New' and 'Open' nav items

* code cleanup

improve tests

* Add new empty dashboard styling

* Fix selenium tests

* Address code review comments

- rename bread_crumb_url => bread_crumb_urls
- fix reporting link in dash (separate PR)
- Use NoItems instead of PromptForItems when searching
- Style fixes

* Use a constant for the landing page url

* Fix tests

fix tests
