+++ src/core_plugins/kibana/public/dashboard/dashboard.html
@@ -6,9 +6,9 @@
       <!-- Title. -->
       <div
         data-transclude-slot="topLeftCorner"
-        class="kuiLocalTitle"
       >
-        <span ng-bind="getDashTitle()"></span>
+        <!-- Breadcrumbs. -->
+        <bread-crumbs title="getDashTitle()" use-links="true" omit-current-page="true"></bread-crumbs>
       </div>
 
       <!-- Search. -->
