+++ src/core_plugins/kibana/public/dashboard/dashboard_constants.js
@@ -1,5 +1,6 @@
 
 export const DashboardConstants = {
   ADD_VISUALIZATION_TO_DASHBOARD_MODE_PARAM: 'addToDashboard',
-  NEW_VISUALIZATION_ID_PARAM: 'addVisualization'
+  NEW_VISUALIZATION_ID_PARAM: 'addVisualization',
+  LANDING_PAGE_URL: '/dashboard'
 };
