--- src/core_plugins/kibana/public/dashboard/top_nav/open.html
@@ -1,2 +1,0 @@
-<div class="kuiLocalDropdownTitle">Open Dashboard</div>
-<saved-object-finder type="dashboards"></saved-object-finder>
