+++ src/ui/public/kbn_top_nav/bread_crumbs/bread_crumbs.js
@@ -1,14 +1,25 @@
 import _ from 'lodash';
 import chrome from 'ui/chrome/chrome';
 import breadCrumbsTemplate from './bread_crumbs.html';
+import { getBreadCrumbUrls } from './bread_crumb_urls';
 import uiModules from 'ui/modules';
 let module = uiModules.get('kibana');
 
-module.directive('breadCrumbs', function () {
+module.directive('breadCrumbs', function ($location) {
   return {
     restrict: 'E',
     scope: {
-      omitCurrentPage: '='
+      omitCurrentPage: '=',
+      /**
+       * Optional title to append at the end of the breadcrumbs
+       * @type {String}
+       */
+      title: '=',
+      /**
+       * If true, makes each breadcrumb a clickable link.
+       * @type {String}
+       */
+      useLinks: '='
     },
     template: breadCrumbsTemplate,
     controller: function ($scope) {
@@ -17,6 +28,11 @@ module.directive('breadCrumbs', function () {
 
       if ($scope.omitCurrentPage === true) {
         $scope.breadcrumbs.pop();
+      }
+
+      if ($scope.useLinks) {
+        const url = '#' + $location.path();
+        $scope.breadCrumbUrls = getBreadCrumbUrls($scope.breadcrumbs, url);
       }
     }
   };
